Can now add new cities

+++ app/src/main/java/com/example/jeppe/weatherapp/models/weatherDataResponse/Wind.java
@@ -10,16 +10,16 @@ import com.google.gson.annotations.SerializedName;
 public class Wind {
     @SerializedName("speed")
     @Expose
-    private Integer speed;
+    private Double speed;
     @SerializedName("deg")
     @Expose
     private Integer deg;
 
-    public Integer getSpeed() {
+    public Double getSpeed() {
         return speed;
     }
 
-    public void setSpeed(Integer speed) {
+    public void setSpeed(Double speed) {
         this.speed = speed;
     }
